pyproject.toml
[project]
name = "oscar"
version = "3.3+u1"
description = "OSCAR Earth system model"
requires-python = ">=3.9"

dependencies = [
  "numpy",
  "xarray",
  "pandas",
  "scipy",
  "matplotlib",
  "pyyaml",
  "netcdf4>=1.7.2",
  "h5netcdf>=1.8.0",
]

[project.scripts]
oscar = "oscar.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
